requirements.txt
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
Flask-WTF==1.1.1
gunicorn==20.1.0
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
SQLAlchemy==2.0.7
Werkzeug==2.2.3
WTForms==3.0.1
